backend/requirements.txt
absl-py==0.7.0
astor==0.7.1
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
numpy==1.16.1
protobuf==3.6.1
pytz==2018.9
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
